second commit

+++ src/main/java/com/tiktok/entity/vo/VideoStatistics.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class VideoStatistics {
-    private Integer allCount;
+    private long allCount;
 
-    private Integer successCount;
+    private long successCount;
 
-    private Integer processCount;
+    private long processCount;
 
-    private Integer passCount;
+    private long passCount;
 }
